First Commit + Google Maps [WIP] + NOAA Integration [WIP] + .gitignore
+ First commit
+ Google Maps is working on iOS but giving errors on Android. Next step
will be to figure out what is causing errors and also add locations,
such as stations and such from NOAA.
+ NOAA Integration has been started. Tested token and API call to
receive stations and this worked. Class created for NOAA objects, not
entirely sure if they are all the same or not though. This needs to be
checked @WaterKH .
+ .gitignore file has been added to get rid of all .dll files, obj and
bin folders, as well as packages.